Add testing
Mocha as testing framework
And a simple test
