+++ package.json
@@ -4,7 +4,7 @@
   "description": "Check if there is a Go upgrade available, and create a pull request if there is.",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "mocha"
   },
   "author": "Stefan 'StefMa' M. (https://StefMa.guru)",
   "homepage": "https://github.com/StefMa/Upgrade-Go-Action",
@@ -20,5 +20,8 @@
     "@actions/core": "^1.10.1",
     "@actions/exec": "^1.1.1",
     "@actions/github": "^5.1.1"
+  },
+  "devDependencies": {
+    "mocha": "^10.2.0"
   }
 }
